Refactor: js파일 모듈로 만듦
