+++ index.html
@@ -7,7 +7,7 @@
     <link rel="stylesheet" href="reset.css">
     <link rel="stylesheet" href="airbnb.css">
     <script defer src="https://use.fontawesome.com/releases/v5.15.2/js/all.js" integrity="sha384-vuFJ2JiSdUpXLKGK+tDteQZBqNlMwAjhZ3TvPaDfN9QmbPb7Q8qUpbSNapQev3YF" crossorigin="anonymous"></script>
-    <script defer src="userBtn.js"></script>
+    <script defer type="module" src="airbnb_module.js"></script>
 </head>
 <body>
     <!-- 코로나 배너 -->
